requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=1.7.0
elasticsearch>=8.0.0
flask>=2.3.0
pyyaml>=6.0.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
